feat:FE/common/ModalInput 구현 완료
