+++ client/src/App.jsx
@@ -1,12 +1,10 @@
 import { Route, Routes } from "react-router-dom";
 import ErrorPage from "./pages/ErrorPage"
-import Test from "./pages/Test"
 
 function App() {
   return (
     <>
       <Routes>
-        <Route path="*" element={<Test />} />
         <Route path="*" element={<ErrorPage />} />
       </Routes>
     </>
